메뉴 좋아요 수와 리뷰 수 수정

+++ src/main/java/com/printf/DDING/entity/MenuLike.java
@@ -18,7 +18,7 @@ public class MenuLike {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long likeNo;
+	private int likeNo;
 
 	@Column(nullable = false)
 	private String menuName;
